% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/build_request.R
\name{build_request}
\alias{build_request}
\alias{build_request.pitchfx}
\title{Method for url objects provided from the payload class}
\usage{
build_request(args, ...)

\method{build_request}{pitchfx}(args = NULL, ...)
}
\arguments{
\item{args}{An object returned by \code{get_payload()}.}

\item{...}{additional arguments}
}
\description{
Method for url objects provided from the payload class
}
\keyword{internal}
