% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payload_methods.R
\name{get_pload.bis_boxscore}
\alias{get_pload.bis_boxscore}
\title{An internal function for bis_boxscore payload.}
\usage{
\method{get_pload}{bis_boxscore}(urlz, ...)
}
\arguments{
\item{urlz}{An urlzect created from a urlz link}

\item{...}{additional arguments}
}
\description{
An internal function for bis_boxscore payload.
}
\keyword{internal}
