% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payload_methods.R
\name{get_pload.game}
\alias{get_pload.game}
\title{An internal function for game payload.}
\usage{
\method{get_pload}{game}(urlz, ...)
}
\arguments{
\item{urlz}{An urlzect created from a urlz link}

\item{...}{additional arguments}
}
\description{
An internal function for game payload.
}
\keyword{internal}
