% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payload_methods.R
\name{get_pload.game_events}
\alias{get_pload.game_events}
\title{An internal function for game_events paylaod.}
\usage{
\method{get_pload}{game_events}(urlz, ...)
}
\arguments{
\item{urlz}{An urlzect created from a urlz link}

\item{...}{additional arguments}
}
\description{
An internal function for game_events paylaod.
}
\keyword{internal}
