% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payload_methods.R
\name{get_pload.inning_hit}
\alias{get_pload.inning_hit}
\title{An internal function for inning_hit payload.}
\usage{
\method{get_pload}{inning_hit}(urlz, ...)
}
\arguments{
\item{urlz}{An urlzect created from a urlz link}

\item{...}{additional arguments}
}
\description{
An internal function for inning_hit payload.
}
\keyword{internal}
