% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payload_methods.R
\name{get_pload.linescore}
\alias{get_pload.linescore}
\title{An internal function for linescore payload.}
\usage{
\method{get_pload}{linescore}(urlz, ...)
}
\arguments{
\item{urlz}{An urlzect created from a urlz link}

\item{...}{additional arguments}
}
\description{
An internal function for linescore payload.
}
\keyword{internal}
