% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statcast.R
\name{statcast_payload}
\alias{statcast_payload}
\title{This is DEV and needs to be transformed into a payload_method}
\usage{
statcast_payload(start, end, player_id = NULL, player_type = NULL)
}
\arguments{
\item{start}{A start date passed as a character in ISO 8601 format. \code{"2017-05-01"}}

\item{end}{An end date passed as a character in ISO 8601 format. \code{"2017-09-01"}}

\item{player_id}{The MLBAM ID of a specific player.}

\item{player_type}{The role of the player as a string. The default is "batter". \code{"batter"} or \code{"pitcher"}}
}
\description{
This is DEV and needs to be transformed into a payload_method
}
\examples{
\dontrun{
statcast_payload(start = "2016-04-06", end = "2016-04-15", player_id = 621043, player_type='batter')

statcast_payload(start = "2016-04-06", end = "2016-04-15", player_id = 592789, player_type='pitcher')

statcast_payload(start = "2016-04-06", end = "2016-04-06")
}
}
