% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payload_methods.R
\name{get_pload}
\alias{get_pload}
\alias{get_pload.inning_all}
\title{Method for payload objects.}
\usage{
get_pload(urlz_obj, ...)

\method{get_pload}{inning_all}(urlz, ...)
}
\arguments{
\item{urlz_obj}{An object returned by \code{get_payload()}.}

\item{...}{additional arguments}
}
\description{
Method for payload objects.
}
\keyword{internal}
