% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/payloadev.R
\name{payload}
\alias{payload}
\alias{payload.statcast}
\alias{payload.pitchfx}
\title{Method for payloadev objects.}
\usage{
payload(args, ...)

\method{payload}{statcast}(args, ...)

\method{payload}{pitchfx}(args, ...)
}
\arguments{
\item{args}{An object returned by \code{get_payload()}.}

\item{...}{additional arguments}
}
\description{
Method for payloadev objects.
}
\keyword{internal}
