% Generated by roxygen2: do not edit by hand
% Please edit documentation in R/statcast.R
\name{statcast_payload}
\alias{statcast_payload}
\title{This is DEV and needs to be transformed into a payload_method}
\usage{
statcast_payload(start, end, player_id = NULL, player_type = NULL)
}
\arguments{
\item{start}{Date of first game for which you want data. Format must be in YYYY-MM-DD format.}

\item{end}{Date of last game for which you want data. Format must be in YYYY-MM-DD format.}

\item{player_id}{The MLBAM ID for the player who's data you want to query.}

\item{player_type}{The player type. Can be 'batter' or 'pitcher'}
}
\description{
This is DEV and needs to be transformed into a payload_method
}
\examples{
\dontrun{
statcast_payload(start = "2016-04-06", end = "2016-04-15", player_id = 621043, player_type='batter')

statcast_payload(start = "2016-04-06", end = "2016-04-15", player_id = 592789, player_type='pitcher')

statcast_payload(start = "2016-04-06", end = "2016-04-06")
}
}
